requirements.txt
absl-py==0.11.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
kiwisolver==1.3.1
lpips==0.1.3
Markdown==3.3.3
matplotlib==3.3.3
ninja==1.10.0.post2
numpy==1.19.3
oauthlib==3.1.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.54.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
